requirements.txt
django
django-allauth
pytz
pyyaml
yaml
djangorestframework
django-rest-passwordreset
psycopg2-binary
ujson
requests
celery
uritemplate
drf-spectacular
pytest
pytest-django
redis